+++ tomcat/src/main/java/org/apache/coyote/http11/controller/LoginController.java
@@ -4,29 +4,50 @@ import java.util.Map;
 import java.util.Optional;
 
 import org.apache.coyote.http11.Http11Processor;
-import org.apache.coyote.http11.request_response.HttpMethod;
 import org.apache.coyote.http11.request_response.HttpStatus;
-import org.apache.coyote.http11.session.Session;
-import org.apache.coyote.http11.session.SessionManager;
 import org.apache.coyote.http11.request_response.request.HttpRequest;
 import org.apache.coyote.http11.request_response.response.HttpResponse;
+import org.apache.coyote.http11.session.Session;
+import org.apache.coyote.http11.session.SessionManager;
+import org.apache.coyote.http11.util.StaticFileReader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
 
-public class LoginController implements Controller {
+public class LoginController extends ServletController {
 
     private static final Logger log = LoggerFactory.getLogger(Http11Processor.class);
 
     @Override
     public boolean supports(HttpRequest request) {
-        return request.getRequestMethod().equals(HttpMethod.POST) && request.getUriPath().equals("/login");
+        return request.getUriPath().equals("/login");
     }
 
     @Override
-    public HttpResponse service(HttpRequest request) {
+    protected HttpResponse doGet(HttpRequest request) {
+        Session session = request.getSession(false);
+        if (session != null) {
+            Object user = session.getAttribute("user");
+            if (user != null) {
+                return HttpResponse.builder()
+                    .status(HttpStatus.Found)
+                    .header("Location", "/index.html")
+                    .body("")
+                    .build();
+            }
+        }
+        String responseBody = new StaticFileReader().readStaticFile("/login.html");
+        return HttpResponse.builder()
+            .status(HttpStatus.OK)
+            .contentType("text/html;charset=utf-8")
+            .body(responseBody)
+            .build();
+    }
+
+    @Override
+    protected HttpResponse doPost(HttpRequest request) {
         Session session = request.getSession(true);
         SessionManager.getInstance().changeSessionId(session);
         try {
