+++ tomcat/src/main/java/org/apache/coyote/http11/controller/StaticResourceController.java
@@ -2,13 +2,12 @@ package org.apache.coyote.http11.controller;
 
 import java.util.Map;
 
-import org.apache.coyote.http11.request_response.HttpMethod;
 import org.apache.coyote.http11.request_response.HttpStatus;
-import org.apache.coyote.http11.util.StaticFileReader;
 import org.apache.coyote.http11.request_response.request.HttpRequest;
 import org.apache.coyote.http11.request_response.response.HttpResponse;
+import org.apache.coyote.http11.util.StaticFileReader;
 
-public class StaticResourceController implements Controller {
+public class StaticResourceController extends ServletController {
 
     private final Map<String, String> contentTypes = Map.of(
         "css", "text/css;charset=utf-8",
@@ -18,23 +17,25 @@ public class StaticResourceController implements Controller {
 
     @Override
     public boolean supports(HttpRequest request) {
-        if (!request.getRequestMethod().equals(HttpMethod.GET)) {
-            return false;
-        }
         String requestUriPath = request.getUriPath();
         return requestUriPath.endsWith(".css") || requestUriPath.endsWith(".html") || requestUriPath.endsWith(".js");
     }
 
     @Override
-    public HttpResponse service(HttpRequest request) {
+    protected HttpResponse doGet(HttpRequest request) {
         String requestUriPath = request.getUriPath();
         int index = requestUriPath.lastIndexOf('.');
         String extension = requestUriPath.substring(index + 1);
         String responseBody = new StaticFileReader().readStaticFile(requestUriPath);
-            return HttpResponse.builder()
-                .status(HttpStatus.OK)
-                .body(responseBody)
-                .contentType(contentTypes.get(extension))
-                .build();
+        return HttpResponse.builder()
+            .status(HttpStatus.OK)
+            .body(responseBody)
+            .contentType(contentTypes.get(extension))
+            .build();
+    }
+
+    @Override
+    protected HttpResponse doPost(HttpRequest request) {
+        throw new IllegalArgumentException("지원하지 않는 http method 입니다");
     }
 }
